Separated logic from just index file, moved endpoints into endpoints, env parsing into env, and session functions into sessions, added call to spotify playlists api
